feat:services,models controllers etc..

+++ src/main/java/ParkingLotAPI/parkingLot/service/VehicleService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import ParkingLotAPI.parkingLot.model.Company;
+import ParkingLotAPI.parkingLot.model.MovementRecord;
 import ParkingLotAPI.parkingLot.model.Vehicle;
 import ParkingLotAPI.parkingLot.model.DTO.VehicleDTO;
 import ParkingLotAPI.parkingLot.repositories.CompanyRepository;
@@ -13,7 +14,6 @@ import ParkingLotAPI.parkingLot.repositories.MovementRecorderRepository;
 import ParkingLotAPI.parkingLot.repositories.VehicleRepository;
 import ParkingLotAPI.parkingLot.service.exceptions.ResourceNotFoundException;
 import ParkingLotAPI.parkingLot.service.exceptions.VehicleNotFoundException;
-import ParkingLotAPI.parkingLot.util.MovementRecord;
 
 @Service
 public class VehicleService {
@@ -23,34 +23,25 @@ public class VehicleService {
 
 	@Autowired
 	private CompanyRepository CompRep;
-	
+
 	@Autowired
 	private MovementRecorderRepository movementRep;
-	
 
 	public Vehicle insert(Vehicle vehicle, String idCompany) {
 
-		Company comp = CompRep.findById(idCompany).orElseThrow(()-> new ResourceNotFoundException(idCompany));
+		Company comp = CompRep.findById(idCompany).orElseThrow(() -> new ResourceNotFoundException("Empresa não encontrada com o ID: " + idCompany));
 
-		if (comp != null) {
-			
-			if(comp.getVehicles().contains(vehicle)) {
-				
-				throw new RuntimeException("Veiculo ja registrado!");
-			}
-			
-			else {
-				
-			vehicle = repository.save(vehicle);
-			comp.getVehicles().add(vehicle);
-			CompRep.save(comp);
-			return vehicle;
-			
-			}
-			
+		if (comp.getVehicles().contains(vehicle)) {
+
+			throw new RuntimeException("Veiculo ja registrado!");
 		}
 
-		throw new ResourceNotFoundException("Empresa não encontrada com o ID: " + idCompany);
+		comp.getVehicles().add(vehicle);
+		vehicle.setCompany(comp);
+		vehicle = repository.save(vehicle);
+		CompRep.save(comp);
+		return vehicle;
+
 	}
 
 	public Vehicle findById(String id) {
@@ -58,16 +49,15 @@ public class VehicleService {
 		System.out.print(vehicle);
 		return vehicle;
 	};
-	
-	
+
 	public Vehicle findByPlaca(String placa) {
-		Vehicle v =  repository.findByPlaca(placa);
-		
-		if(v  == null) {
-			 throw new VehicleNotFoundException("Vehicle not found");
+		Vehicle v = repository.findByPlaca(placa);
+
+		if (v == null) {
+			throw new VehicleNotFoundException("Vehicle not found");
 
 		}
-		
+
 		return v;
 	}
 
@@ -76,10 +66,17 @@ public class VehicleService {
 	}
 
 	public void deleteById(String id) {
-		Vehicle vehicle = repository.findById(id).orElseThrow();
+		
+		Vehicle vehicle = repository.findById(id).orElseThrow(()->  new VehicleNotFoundException(id));
 		MovementRecord mr = vehicle.getMovementRecord();
+		
+		if(mr==null) {
+			throw new ResourceNotFoundException("Associated Movement Record not found");
+		}
+		
 		movementRep.delete(mr);
 		repository.delete(vehicle);
+		CompRep.save(vehicle.getCompany());
 	}
 
 	public Vehicle fromDto(VehicleDTO objdto) {
